feat(backend deployment): adding backend url string, github secret
Adding .env.local variable to the frontend folder. Deployed the image to cloud run on gcp. Also I
updated the scripts , dependencies in order  for the container image to be properly deployed, had
issues with renning the typscript compiler and  so on. H Updated github build script in order to
make the  secret... secret. Small frontend change to display the result. Time to eat :)